techpack: display: protect setting doze status by allow_bl_update
In the current implementation the doze status command might set the backlight
of the panel before receiving a frame from the driver during the first commit.
There is chance for garbage content to be shown onto the panel.
This change imposes a condition to delay the fod hbm command until the
first frame is received from the HW.